finish dataframe
